refactor: extract monthpicker logic
